+++ tests/extension.test.ts
@@ -126,6 +126,7 @@ describe('Extension', () => {
 
       expect(getExtensionSettings()).toEqual({
         autoEnable: true,
+        coverageFormatter: 'DefaultFormatter',
         debugCodeLens: {
           enabled: true,
           showWhenTestStateIn: [TestState.Fail, TestState.Unknown],
@@ -139,6 +140,7 @@ describe('Extension', () => {
         rootPath: '<rootDir>',
         runAllTestsFirst: true,
         showCoverageOnLoad: false,
+        debugMode: false,
       })
     })
   })